mcp-provider-delegator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-provider-delegator"
version = "0.1.0"
description = "MCP server for delegating agents to AI providers (Codex, Gemini) with fallback support"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0.1",
]

[project.scripts]
mcp-provider-delegator = "mcp_provider_delegator.server:run"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_provider_delegator"]
